+++ src/Pim/Bundle/CatalogBundle/Form/Type/ProductEditType.php
@@ -9,7 +9,6 @@ use Symfony\Component\Form\AbstractType;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
 use Pim\Bundle\CatalogBundle\Form\View\ProductFormView;
 use Pim\Bundle\CatalogBundle\Form\Subscriber\BindAssociationTargetsSubscriber;
-use Pim\Bundle\CatalogBundle\Form\Subscriber\IgnoreMissingFieldDataSubscriber;
 
 /**
  * Product edit form type
@@ -92,8 +91,7 @@ class ProductEditType extends AbstractType
                     'mapped'   => true,
                     'multiple' => true,
                 )
-            )
-            ->addEventSubscriber(new IgnoreMissingFieldDataSubscriber());
+            );
     }
 
     /**